backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.38.0
python-multipart==0.0.20
requests==2.32.5
beautifulsoup4==4.14.2
pydantic==2.12.4
pydantic[email]==2.12.4
email-validator>=2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-dotenv==1.2.1
gunicorn==21.2.0
supabase==2.11.0
duckduckgo-search==7.3.2
selenium==4.29.0
webdriver-manager==4.0.2
google-generativeai==0.8.3
